refactor(opencode): split user message helpers
